requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.9.0
python-multipart>=0.0.6
google-cloud-storage==2.14.0
google-cloud-discoveryengine>=0.15.0
pandas>=2.2.3
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
beautifulsoup4==4.12.2
lxml>=5.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
google-auth>=2.23.0

